INS-1644 fixing/renaming/spliting files in drop package
