+++ ledger/storage/storage_test.go
@@ -75,7 +75,7 @@ func (s *storageSuite) BeforeTest(suiteName, testName string) {
 
 	s.objectStorage = storage.NewObjectStorage()
 
-	dropStorage := jetdrop.NewDropStorageDB()
+	dropStorage := jetdrop.NewStorageDB()
 	s.dropAccessor = dropStorage
 	s.dropModifier = dropStorage
 	s.pulseTracker = storage.NewPulseTracker()
@@ -258,7 +258,7 @@ func TestDB_Close(t *testing.T) {
 	jetID := testutils.RandomJet()
 
 	os := storage.NewObjectStorage()
-	ds := jetdrop.NewDropStorageDB()
+	ds := jetdrop.NewStorageDB()
 
 	cm := &component.Manager{}
 	cm.Inject(
